+++ Models/Student.cs
@@ -12,6 +12,6 @@ namespace Asal.StudentManagementSystem.Models
 
         public int Age { get; set; }
 
-        public int Grade { get; set; }
+        public double Grade { get; set; }
     }
 }
